requirements.txt
fastapi
uvicorn
pdfplumber
openai
slowapi
python-multipart
python-dotenv
